[ci] Improve parallelism of yarn test
This PR adds parallelism similar to our existing circleci setup for
running yarn tests with the various test params. It does this by first
chunking tests into groups of `$parallelism`, then spawning a job for
each.

Effectively this means that the job will spawn an additional (where `n`
is the number of test params)

`n` number of jobs to chunk tests
`n * $parallelism` number of jobs to run tests in parallel

for a total of `n + n + (n * $parallelism)` jobs. This does mean the
GitHub UI at the bottom of each PR gets longer and unfortunately it's
not sorted in any way as far as I can tell. But if something goes wrong
it should still be easy to find out what the problem is.

Note that the xplat tests don't include react-dom so we pass the
`--passWithNoTests` flag in case that chunk only receives react-dom
tests. We could be smarter about this in the future.

The PR also changes the `ci` argument for jest-cli to be an enum instead
so the tests use all available workers in GitHub actions. This will have
to live around for a bit until we can fully migrate off of circleci.

ghstack-source-id: d4cc3e3f6d8bc9520c66472fd2a87722cedda2c8
Pull Request resolved: https://github.com/facebook/react/pull/30033